feat!: Relaxed stack initialization
Commands that create patches will now auto-initialize the stack if it is
not already initialized. This includes `new`, `import`, `pick`, and
`uncommit`.

For each of these commands, the user has given a strong affirmative signal
that they want to create patches, so now StGit does the stack init step and
achieves the user's desired outcome instead of failing, forcing the user to
perform the rote/trivial `stg init`, and then forcing the user to repeat
their patch-creating command.

For most other non patch creating commands, if the stack is not
initialized, an empty Stack instance is now used. Commands such as `series`
and `show` can thus succeed with meaningful output. Other commands such as
`top`, `prev`, `next`, `push`, and `pop` still fail, but in the manner
expected with an empty stack as opposed to with an uninitialized stack
error.

The rationale: because patch creating commands now auto-initialize the
stack, there is more value to the user if other commands inform about the
absence of patches rather than the stack being uninitialized.

Commands that work with stack history such as `log`, `undo`, `redo`,
`repair`, and `reset` still fail the same way: with a stack uninitialized
error.

N.B. any commands allowing an uninitialized stack must not attempt to
perform a stack transaction. A panic will happen at runtime if such a
transaction is attempted.

*This is a breaking change. Users or scripts expecting stack not
initialized errors in various contexts will now see either successful
outcomes or errors stemming from the stack being empty.*

Refs: #238

+++ src/cmd/clean.rs
@@ -10,7 +10,7 @@ use crate::{
     commit::CommitExtended,
     patchname::PatchName,
     repo::RepositoryExtended,
-    stack::{Stack, StackStateAccess},
+    stack::{InitializationPolicy, Stack, StackStateAccess},
     stupid::Stupid,
 };
 
@@ -47,7 +47,7 @@ fn make() -> clap::Command {
 
 fn run(matches: &ArgMatches) -> Result<()> {
     let repo = git2::Repository::open_from_env()?;
-    let stack = Stack::from_branch(&repo, None)?;
+    let stack = Stack::from_branch(&repo, None, InitializationPolicy::AllowUninitialized)?;
     stack.check_head_top_mismatch()?;
     repo.check_repository_state()?;
 
@@ -85,17 +85,19 @@ fn run(matches: &ArgMatches) -> Result<()> {
         }
     }
 
-    stack
-        .setup_transaction()
-        .allow_conflicts(true)
-        .use_index_and_worktree(false)
-        .with_output_stream(get_color_stdout(matches))
-        .transact(|trans| {
-            let to_push = trans.delete_patches(|pn| to_delete.contains(pn))?;
-            trans.push_patches(&to_push, false)?;
-            Ok(())
-        })
-        .execute("delete")?;
+    if !to_delete.is_empty() {
+        stack
+            .setup_transaction()
+            .allow_conflicts(true)
+            .use_index_and_worktree(false)
+            .with_output_stream(get_color_stdout(matches))
+            .transact(|trans| {
+                let to_push = trans.delete_patches(|pn| to_delete.contains(pn))?;
+                trans.push_patches(&to_push, false)?;
+                Ok(())
+            })
+            .execute("delete")?;
+    }
 
     Ok(())
 }
